feat: allow references to the selected files during shell execution
